Исправления по замечаниям наставника
