+++ js/main.js
@@ -2,9 +2,11 @@ import { photos } from './data';
 import { renderPhotos } from './picture';
 import { generateBigPictureModal } from './gallery';
 import { uploadImg } from './img-upload-form';
-import './img-editing';
-import './img-effect';
+import { configureImgResizing } from './img-resizing';
+import { getImgEffect } from './img-effect';
 
 renderPhotos(photos);
 generateBigPictureModal();
 uploadImg();
+configureImgResizing();
+getImgEffect();
